Add finished task

+++ src/main/java/Calculator.java
@@ -2,40 +2,40 @@
 
 public class Calculator {
 
-    private int result;    // Variable for the result
+    private double result;    // Variable for the result
 
     public void reset() {  // Resets the result variable
         result = 0;
     }
 
-    public int getResult() {
+    public double getResult() {
         return result;
     }
 
-    public void add(int n) {
+    public void add(double n) {
         result = result + n;
     }
 
-    public void subtract(int n) {
+    public void subtract(double n) {
         result = result - n;
     }
 
-    public void multiply(int n) {
+    public void multiply(double n) {
         result = result * n;
     }
 
-    public void divide(int n) {
+    public void divide(double n) {
         if (n == 0) throw new ArithmeticException("Cannot divide by zero");
         result = result / n;
     }
 
-    public void square(int n) {
-        result = (int) Math.pow(n, 2);
+    public void square(double n) {
+        result = (double) Math.pow(n, 2);
     }
 
-    public void squareRoot(int n) {
+    public void squareRoot(double n) {
         if (n < 0) throw new ArithmeticException("Cannot get square root of negative number");
-        result = (int) Math.sqrt(n);
+        result = (double) Math.sqrt(n);
     }
 
     public void powerOn() {
